requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0

# FAISS library (choose one based on your needs)
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU support

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
kaleido>=0.2.1  # For plotly static image export

# Machine learning utilities
scikit-learn>=1.0.0

# Performance monitoring
psutil>=5.8.0

# Data processing
h5py>=3.6.0
tqdm>=4.62.0

# CLI utilities
click>=8.0.0

# Optional: Jupyter notebook support
# jupyter>=1.0.0
# ipywidgets>=7.6.0

# Development dependencies (optional)
# pytest>=6.0.0
# black>=21.0.0
# flake8>=3.9.0